Renamed misc to misc funcs for clarity
